feat: switch money flow to kopecks-only storage semantics

+++ src/routes/sellerRoutes.ts
@@ -325,7 +325,7 @@ const sellerMediaUrlSchema = z.string().refine((value) => {
 
 const sellerVariantMutationSchema = z.object({
   sku: z.string().min(3).optional(),
-  price: z.number().int().positive().optional(),
+  price: z.number().positive().optional(),
   color: z.string().min(2).optional(),
   variantLabel: z.string().min(1).max(120).optional(),
   variantSize: z.string().min(1).max(64).optional(),
